Trello-VYvzod96. Refactoring
