+++ client/store/auth/reducerMethods.test.js
@@ -6,10 +6,6 @@ import {
     authLogInCookieFinishedHandler,
     authLogInCookieFailedHandler,
     logOutHandler,
-    getUserProfileStartedHandler,
-    getUserProfileFinishedHandler,
-    getUserProfileFailedHandler,
-    changeProfileLanguageFinishedHandler,
     createNewAccountStartedHandler,
     createNewAccountFinishedHandler,
     createNewAccountFailedHandler
@@ -539,320 +535,6 @@ describe('reducers', () => {
                     },
                     profile: {
                         name: 'User Name'
-                    }
-                });
-            });
-        });
-
-        describe('"getUserProfileStartedHandler" method', () => {
-            test('it should be defined', () => {
-                expect(typeof getUserProfileStartedHandler).toBe('function');
-            });
-
-            test('it should return correct new state', () => {
-                let newState;
-                let state = {
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profile: {}
-                };
-
-                newState = getUserProfileStartedHandler(state);
-
-                expect(newState).toEqual({
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: true,
-                        error: ''
-                    },
-                    profile: {}
-                });
-            });
-
-            test('it should not mutate original state', () => {
-                let state = {
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profile: {}
-                };
-
-                getUserProfileStartedHandler(state);
-
-                expect(state).toEqual({
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profile: {}
-                });
-            });
-        });
-
-        describe('"getUserProfileFinishedHandler" method', () => {
-            test('it should be defined', () => {
-                expect(typeof getUserProfileFinishedHandler).toBe('function');
-            });
-
-            test('it should return correct new state', () => {
-                let newState;
-                let profileData = {
-                    name: 'User Name',
-                    lang: 'ru'
-                };
-                let state = {
-                    language: 'en',
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: true,
-                        error: ''
-                    },
-                    profile: {}
-                };
-
-                newState = getUserProfileFinishedHandler(state, profileData);
-
-                expect(newState).toEqual({
-                    language: 'ru',
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: true,
-                    profileFetching: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profile: {
-                        lang: 'ru',
-                        name: 'User Name'
-                    }
-                });
-            });
-
-            test('it should not mutate original state', () => {
-                let profileData = {
-                    name: 'User Name'
-                };
-                let state = {
-                    language: 'en',
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: true,
-                        error: ''
-                    },
-                    profile: {}
-                };
-
-                getUserProfileFinishedHandler(state, profileData);
-
-                expect(state).toEqual({
-                    language: 'en',
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: true,
-                        error: ''
-                    },
-                    profile: {}
-                });
-            });
-        });
-
-        describe('"getUserProfileFailedHandler" method', () => {
-            test('it should be defined', () => {
-                expect(typeof getUserProfileFailedHandler).toBe('function');
-            });
-
-            test('it should return correct new state', () => {
-                let newState;
-                let state = {
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: true,
-                        error: ''
-                    },
-                    profile: {}
-                };
-
-                newState = getUserProfileFailedHandler(state, 'Some Error');
-
-                expect(newState).toEqual({
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: false,
-                        error: 'Some Error'
-                    },
-                    profile: {}
-                });
-            });
-
-            test('it should not mutate original state', () => {
-                let state = {
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: true,
-                        error: ''
-                    },
-                    profile: {}
-                };
-
-                getUserProfileFailedHandler(state, 'Some Error');
-
-                expect(state).toEqual({
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: false,
-                    profileFetching: {
-                        inProgress: true,
-                        error: ''
-                    },
-                    profile: {}
-                });
-            });
-        });
-
-        describe('"changeProfileLanguageFinishedHandler" method', () => {
-            test('it should be defined', () => {
-                expect(typeof changeProfileLanguageFinishedHandler).toBe('function');
-            });
-
-            test('it should return correct new state', () => {
-                let newState;
-                let state = {
-                    language: 'en',
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: true,
-                    profileFetching: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profile: {
-                        username: 'username',
-                        activeProject: null
-                    }
-                };
-
-                newState = changeProfileLanguageFinishedHandler(state, 'ru');
-
-                expect(newState).toEqual({
-                    language: 'ru',
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: true,
-                    profileFetching: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profile: {
-                        username: 'username',
-                        activeProject: null
-                    }
-                });
-            });
-
-            test('it should not mutate original state', () => {
-                let state = {
-                    language: 'en',
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: true,
-                    profileFetching: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profile: {
-                        username: 'username',
-                        activeProject: null
-                    }
-                };
-
-                changeProfileLanguageFinishedHandler(state, 'ru');
-
-                expect(state).toEqual({
-                    language: 'en',
-                    authorized: true,
-                    authorization: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profileFetched: true,
-                    profileFetching: {
-                        inProgress: false,
-                        error: ''
-                    },
-                    profile: {
-                        username: 'username',
-                        activeProject: null
                     }
                 });
             });
