+++ client/store/profile/selectors.js
@@ -1,0 +1,1 @@
+export const getLanguage = (state) => (state.profile.data.lang);
