+++ client/store/reducers.js
@@ -2,10 +2,12 @@ import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
 
 import authReducer from './auth/reducer';
+import profileReducer from './profile/reducer';
 import projectsReducer from './projects/reducer';
 
 export default combineReducers({
     routing: routerReducer,
-    user: authReducer,
-    projects: projectsReducer
+    profile: profileReducer,
+    projects: projectsReducer,
+    user: authReducer
 });
