+++ client/translate/translate.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import { getLanguage } from 'store/profile/selectors';
 
 export default function translate(languages) {
     return Component => {
@@ -19,7 +20,7 @@ export default function translate(languages) {
 
         function mapStateToProps(state) {
             return {
-                currentLanguage: state.user.language
+                currentLanguage: getLanguage(state)
             };
         }
